Initial implementation
This is the first implementation of a coomunity board.

This board is a NUCLEO64-F411RE with the ST-LINK part cut off.
As a result of the missing ST_Link part we have no external xtal anymore.
Additional components have been added to the board to have a HSE again.
How to achieve this can all be read in the UM1724 manual on page 24.
The manual can be found on http://www.st.com and search for NUCLEO-F411RE.
The all power supply is gone I have cut a USB cable to connect PA11->DM, PA12->DP and use the USB power wires too.
Don't forget to set jumper JP5 to use E5V and put Boot0 to Vss to put the device in DFU mode.
Use the STDFUTester application of ST to erase and burn the image.

The images build will work of course on a stock board not modified as described above.

Signed-off-by: Peter Wessel <piwi1263@gmail.com>